requirements_production.txt
pytest
pylint
ipdb
ipykernel
jupyter
jupyterlab
ujson

numpy
pandas
matplotlib
seaborn
scikit-learn

# Tensoflow packages
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# Google Cloud
google-cloud-bigquery
google-cloud-storage

mlflow==2.1.1
prefect==2.14.9

# API
fastapi==0.108.0
# # pytz
uvicorn

# Prettify
tqdm
